fix: move connect to top level
